bun_core: move debug module to its own file
